# services/telegram-bot/openapi.yaml
openapi: 3.0.0
info:
  description: Сервис для взаимодействия с телеграммом
  title: telegram-bot
  version: 1.0.0
servers:
- url: /
paths:
  /telegram-bot/ping:
    get:
      operationId: getGreeting
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getGreeting_200_response'
          description: Successful response
      summary: Returns a greeting and the current time
  /telegram-bot/internal/send:
    post:
      operationId: sendMessage
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/sendMessage_request'
        required: true
      responses:
        "204":
          description: Ok
        "400":
          description: Bad Request
        "404":
          description: Not Found
      summary: Send message to the user
components:
  schemas:
    sendMessage_request:
      properties:
        chat_id:
          title: chat_id
          type: integer
        message_text:
          minLength: 1
          title: message_text
          type: string
      required:
      - chat_id
      - message_text
      title: sendMessage_request
      type: object
    getGreeting_200_response_tg_bot_thing:
      example:
        test_prop: test_prop
      properties:
        test_prop:
          title: test_prop
          type: string
      title: getGreeting_200_response_tg_bot_thing
      type: object
    getGreeting_200_response:
      example:
        currentTime: 2023-10-10T14:48:00Z
        tg_bot_thing:
          test_prop: test_prop
        message: "Hello, World!"
      properties:
        message:
          example: "Hello, World!"
          title: message
          type: string
        currentTime:
          example: 2023-10-10T14:48:00Z
          format: date-time
          title: currentTime
          type: string
        tg_bot_thing:
          $ref: '#/components/schemas/getGreeting_200_response_tg_bot_thing'
      title: getGreeting_200_response
      type: object